Add check for inkscape in stroke to fill script
Inkscape is a godsend for this script so...

+++ stroke_to_fill.sh
@@ -1,7 +1,11 @@
+if ! [ -x "$(command -v inkscape)" ]; then
+	echo 'Error: inkscape is required, but is not installed.' >&2
+	exit 1
+fi
+
 rm -rf filled
 cp -r stroked/ filled/
 
-for f in filled/*
-do
+for f in filled/*; do
 	inkscape -g --verb=EditSelectAll --verb=SelectionUnGroup --verb=StrokeToPath --verb=EditDeselect --verb=FileSave --verb=FileQuit $f
 done
